move(MemberRepository): MemberBioInfo 엔티티 member.domain 패키지로 이동
