+++ src/test/java/soma/ghostrunner/domain/auth/application/AuthServiceTest.java
@@ -14,7 +14,7 @@ import soma.ghostrunner.domain.auth.exception.TokenTheftException;
 import soma.ghostrunner.domain.auth.resolver.impl.FirebaseUidResolver;
 import soma.ghostrunner.domain.member.InvalidMemberException;
 import soma.ghostrunner.domain.member.domain.Member;
-import soma.ghostrunner.domain.member.MemberRepository;
+import soma.ghostrunner.domain.member.dao.MemberRepository;
 import soma.ghostrunner.domain.member.application.MemberService;
 import soma.ghostrunner.domain.member.dao.MemberAuthInfoRepository;
 import soma.ghostrunner.domain.member.domain.MemberAuthInfo;
