+++ src/test/java/soma/ghostrunner/domain/member/dao/MemberAuthInfoRepositoryTest.java
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import soma.ghostrunner.IntegrationTestSupport;
 import soma.ghostrunner.domain.member.domain.Member;
-import soma.ghostrunner.domain.member.MemberRepository;
 import soma.ghostrunner.domain.member.domain.MemberAuthInfo;
 
 import java.util.Optional;
