+++ src/test/java/soma/ghostrunner/learning/JpaModifyingTest.java
@@ -10,7 +10,7 @@ import soma.ghostrunner.domain.course.domain.Course;
 import soma.ghostrunner.domain.course.domain.CourseProfile;
 import soma.ghostrunner.domain.course.domain.StartPoint;
 import soma.ghostrunner.domain.member.domain.Member;
-import soma.ghostrunner.domain.member.MemberRepository;
+import soma.ghostrunner.domain.member.dao.MemberRepository;
 import soma.ghostrunner.domain.running.dao.RunningRepository;
 import soma.ghostrunner.domain.running.domain.Running;
 import soma.ghostrunner.domain.running.domain.RunningMode;
